requirements.txt
requests
pygame
customtkinter
Pillow
opencv-python